Scipts/Scripts for compiling the _Scientific results_ and _Technologies & tools_ catalogues/requirements.txt
pandas
tqdm
google
protobuf
python-dotenv
requests
beautifulsoup4
google-generativeai
lxml
openpyxl
